+++ demo/app/controllers/preview_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require "rails/application_controller"
+
+class PreviewController < ViewComponentsController # :nodoc:
+  helper Primer::ViewHelper
+end